+++ src/main/java/io/github/sfseeger/manaweave_and_runes/core/init/RitualInit.java
@@ -32,4 +32,5 @@ public class RitualInit {
     public static final Supplier<ShatteringRiteRitual> SHATTERING_RITE_RITUAL =
             RITUALS.register("shattering_rite_ritual", ShatteringRiteRitual::new);
 
+    public static final Supplier<SmiteRitual> SMITE_RITUAL = RITUALS.register("smite_ritual", SmiteRitual::new);
 }