+++ src/main/java/io/github/sfseeger/manaweave_and_runes/datagen/server/tags/MRBlockTagsProvider.java
@@ -21,6 +21,8 @@ public class MRBlockTagsProvider extends BlockTagsProvider {
     @Override
     protected void addTags(HolderLookup.Provider provider) {
         this.tag(BlockTags.MINEABLE_WITH_PICKAXE)
+                .add(MRBlockInit.TANZANITE_ORE.get())
+                .add(MRBlockInit.DEEPSLATE_TANZANITE_ORE.get())
                 .add(MRBlockInit.NOVICE_MANA_COLLECTOR.get())
                 .add(MRBlockInit.MANA_STORAGE_BLOCK.get())
                 .add(MRBlockInit.NOVICE_MANA_CONCENTRATOR_BLOCK.get())
@@ -40,7 +42,8 @@ public class MRBlockTagsProvider extends BlockTagsProvider {
                 .add(MRBlockInit.ORDER_MANA_INFUSED_ROCK_BLOCK.get())
                 .add(MRBlockInit.SOUL_MANA_INFUSED_ROCK_BLOCK.get())
                 .add(MRBlockInit.VOID_MANA_INFUSED_ROCK_BLOCK.get())
-                .add(MRBlockInit.RUNE_PEDESTAL_BLOCK.get());
+                .add(MRBlockInit.RUNE_PEDESTAL_BLOCK.get())
+        ;
 
         this.tag(BlockTags.NEEDS_STONE_TOOL)
                 .add(MRBlockInit.RUNE_BLOCK.get())
@@ -52,7 +55,8 @@ public class MRBlockTagsProvider extends BlockTagsProvider {
                 .add(MRBlockInit.ORDER_MANA_INFUSED_ROCK_BLOCK.get())
                 .add(MRBlockInit.SOUL_MANA_INFUSED_ROCK_BLOCK.get())
                 .add(MRBlockInit.VOID_MANA_INFUSED_ROCK_BLOCK.get())
-                .add(MRBlockInit.RUNE_PEDESTAL_BLOCK.get());
+                .add(MRBlockInit.RUNE_PEDESTAL_BLOCK.get())
+        ;
 
         this.tag(BlockTags.NEEDS_IRON_TOOL)
                 .add(MRBlockInit.NOVICE_MANA_COLLECTOR.get())
@@ -64,11 +68,13 @@ public class MRBlockTagsProvider extends BlockTagsProvider {
                 .add(MRBlockInit.MASTER_RITUAL_ANCHOR_BLOCK.get())
                 .add(MRBlockInit.ASCENDED_RITUAL_ANCHOR_BLOCK.get())
                 .add(MRBlockInit.MANA_TRANSMITTER_BLOCK.get())
-                .add(MRBlockInit.RUNEWROUGHT_BENCH_BLOCK.get());
+                .add(MRBlockInit.RUNEWROUGHT_BENCH_BLOCK.get())
+        ;
 
         this.tag(BlockTags.NEEDS_DIAMOND_TOOL)
                 .add(MRBlockInit.TANZANITE_ORE.get())
-                .add(MRBlockInit.DEEPSLATE_TANZANITE_ORE.get());
+                .add(MRBlockInit.DEEPSLATE_TANZANITE_ORE.get())
+        ;
 
         this.tag(MRTagInit.MANA_INFUSED_BLOCK)
                 .add(MRBlockInit.FIRE_MANA_INFUSED_ROCK_BLOCK.get())