+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## 1.21.1-0.1.1
+## 1.21.1-0.1.2
+- Add smite ritual
+- Fix Tanzanite Ore not dropping when mined
+- Make template duplication more expensive
 
+## 1.21.1-0.1.1
 - Updated Runecarver Texture
 - Added some patchouli documentation
 - Added missing mana textures